+++ command.h
@@ -1,8 +1,15 @@
 #ifndef COMMAND_H_
 #define COMMAND_H_
 
+#include "./net.h"
+
+#define MAX_CMD_SIZE 4096
+#define MAX_KEY_SIZE 256
 #define DEFAULT_REPLY_LINES 16
 #define DEFAULT_REPLY_SIZE 256
+
+#define LAST_LINE(r) (r->i > 0 ? r->lines[r->i - 1] : NULL)
+#define LAST_LINE2(r) (r.i > 0 ? r.lines[r.i - 1] : NULL)
 
 #define INIT_REPLY(r) do {\
   r->size = DEFAULT_REPLY_LINES;\
@@ -37,6 +44,9 @@
   reply->types[reply->i] = NIL;\
   reply->i++;\
 } while (0)
+
+typedef enum { STRING, INTEGER, RAW, ERR, NIL } ReplyType;
+typedef struct { int size, i, *sizes; char **lines; ReplyType *types; } Reply;
 
 int command(Conn *, const char *, Reply *);
 int pipeline(Conn *, const char *, Reply *, int);
