+++ command_raw.h
@@ -1,6 +1,8 @@
 #ifndef COMMAND_RAW_H_
 #define COMMAND_RAW_H_
 
+#include "./net.h"
+
 int commandWithRawData(Conn *, const void *, Reply *, int);
 
 #endif // COMMAND_RAW_H_
