+++ generic.h
@@ -6,12 +6,6 @@
 
 #define MY_OK_CODE 0
 #define MY_ERR_CODE -1
-#define ANY_NODE_OK -2
-#define MAX_HOST_SIZE 256
-#define MAX_PORT_SIZE 8
-#define MAX_CMD_SIZE 4096
-#define MAX_KEY_SIZE 256
-#define CLUSTER_SLOT_SIZE 16384
 
 #define ASSERT(ret) do {\
   if (ret == MY_ERR_CODE) exit(1);\
@@ -30,16 +24,5 @@
     exit(1);\
   }\
 } while (0)
-
-#define LAST_LINE(r) (r->i > 0 ? r->lines[r->i - 1] : NULL)
-#define LAST_LINE2(r) (r.i > 0 ? r.lines[r.i - 1] : NULL)
-#define FIND_CONN(c, i) (c->nodes[c->slots[i]])
-#define FIND_CONN2(c, i) (c.nodes[c.slots[i]])
-
-typedef struct { char host[MAX_HOST_SIZE], port[MAX_PORT_SIZE]; } HostPort;
-typedef struct { FILE *fw, *fr; HostPort addr; } Conn;
-typedef struct { int size, i; Conn **nodes; int slots[CLUSTER_SLOT_SIZE]; } Cluster;
-typedef enum { STRING, INTEGER, RAW, ERR, NIL } ReplyType;
-typedef struct { int size, i, *sizes; char **lines; ReplyType *types; } Reply;
 
 #endif // GENERIC_H_
