+++ net.h
@@ -1,6 +1,12 @@
 #ifndef NET_H_
 #define NET_H_
 
+#define MAX_HOST_SIZE 256
+#define MAX_PORT_SIZE 8
+
+typedef struct { char host[MAX_HOST_SIZE], port[MAX_PORT_SIZE]; } HostPort;
+typedef struct { FILE *fw, *fr; HostPort addr; } Conn;
+
 int createConnection(Conn *);
 int createConnectionFromStr(const char *, Conn *);
 int reconnect(Conn *);
